recommendation_api/requirements.txt
azure-functions
numpy==1.26.4
pandas==2.2.2
scikit-surprise==1.1.3
scipy==1.11.4
azure-storage-blob==12.20.0

